hidive

Module rescue

Enums

SearchOption 🔒

Functions

collect_loci 🔒

This function collects the loci from the mappings.

create_fetch_definition 🔒

This function creates a fetch definition.

detect_relevant_loci 🔒

This function detects the relevant loci in the reference genome.

filter_and_collect_records 🔒

This function processes all or unmapped reads.

finalize_progress 🔒

This function finalizes the progress.

initialize_aligner 🔒

This function initializes the aligner.

is_valid_read 🔒

This function checks if a read is valid.

map_sequences 🔒

This function maps the sequences to the reference genome using the aligner.

merge_overlapping_intervals 🔒

This function merges overlapping intervals.

populate_fetches 🔒

This function will populate the fetches with the input loci or the detected relevant loci. Function should only be called if the search option is set to ‘Contig’ or ‘ContigAndInterval’.

populate_interval_trees 🔒

This function populates the interval trees with the loci.

prepare_fetches 🔒

This function prepares the fetches for searching. If the search option is ‘Contig’, it will extract the unique contigs from the fetches and add a dummy interval. If the search option is ‘ContigAndInterval’, it will return the fetches as is. In both cases, the fetches will be returned as a vector of contigs and intervals.

print_fetches_info 🔒

This function will print the fetches’ info.

retrieve_records 🔒

This function retrieves the records from the reader.

start

update_found_progress 🔒

This function updates the found progress.

update_processed_progress 🔒

This function updates the processed progress.

write_paired_fastq_records 🔒