hidive::rescue
Function prepare_fetches

fn prepare_fetches(
    fetches: &[(String, Interval<i32>)],
    search_option: &SearchOption,
) -> Vec<(String, Interval<i32>)>

This function prepares the fetches for searching. If the search option is ‘Contig’, it will extract the unique contigs from the fetches and add a dummy interval. If the search option is ‘ContigAndInterval’, it will return the fetches as is. In both cases, the fetches will be returned as a vector of contigs and intervals.

Arguments

fetches - The fetches to prepare.

search_option - Whether to search all reads or not.

Returns

A vector of fetches to search.

Panics

If the contig is not valid.